flask_app/requirements.txt
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
Flask-BasicAuth==0.2.0
Flask-Caching==2.3.0
python-dotenv==1.0.1
pymemcache==4.0.0
gunicorn==22.0.0
